+++ Frenetik.iRacingApiWrapper/Models/MemberAward.cs
